third_party/blink/web_tests/external/wpt/css/mediaqueries/aspect-ratio-serialization-expected.txt

This is a testharness.js-based test.
FAIL <ratio> serializes with spaces around the integer. assert_equals: expected "(aspect-ratio: 1 / 3)" but got "(aspect-ratio: 1/3)"
Harness: the test ran to completion.
